tunefs.ocfs2: Fix a substring bug in op_set_label.
If tunefs was trying to set the label of a filesystem to a substring of
the current label, it would consider the label already set, because it
wasn't checking the string terminator.

For example, a filesystem has the label "server12".  tunefs is run with
the label "server1" (that is, 'tunefs -L server1').  tunefs will compare
the two strings, but only to the length of the new label.  Since the new
label is the same as the start of the old label, tunefs says "label
already set" and does nothing.  This is wrong.

The fix is to compare one byte more than the length of the new label.
This includes the string terminating NUL character.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>