debugfs.ocfs2: Swap blocks we're dumping from the journal.
The 'logdump' command reads blocks from the journal and tries to guess
what kind of block they are.  If it finds a known block type, it then
dumps the block.

However, there is a bug.  The dump_*() functions of degugfs.ocfs2 assume
the block has already been brought to cpu-endian.  The other callers
(regular debugfs commands) swap the blocks.  So we need to swap the
blocks in dump_jbd_metadata() before we call the dump_*() functions.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>