Ocfs2-tools: Correct fs_locks dump for debugfs.ocfs2
Noticed that dump_fs_locks only dumps half of the all lockres, it was due to
the fault of dump_version_two() func, this func seems has been added early for
the sake of keeping compatible for version two. and it was used to dump 9 more
records which were new in version 2. but unfortuately, the author seems wrongly
read the '\t' of last record of each lock entry. this patch correct such a reading
mistake.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>