ocfs2console: Handle an invalid DISPLAY gracefully with modern pygtk.
Current pygtk treats an invalid DISPLAY as a WARNING.  This means that
pygtk initialization tries to continue even after gtk_init() has failed.
All sorts of fun ensues.  To prevent this, we turn the warning into an
error for the duration of gtk initialization.  This is really a
workaround for pygtk.

Originally reported as Novell bugzilla #448523.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: coly li <coyli@novell.com>