debugfs.ocfs2: Dump refcount trees
debugfs.ocfs2 gains the 'refcount' command.  This will dump a refcount
tree including its extents.  If passed an inode specification, it will
dump the refcount tree associated with that inode.  If it is passed the
refcount block directly (surrounded by angle brackets, of course), it
will just read the refcount block.

Signed-off-by: Joel Becker <joel.becker@oracle.com>