ocfs2_controld: Add a couple helpers for functions that retry operations
Add a helper function to sleep for a certain number of milliseconds, and
add another function that will increase a count, printing an error
whenever that count is a power of two.  eg:

    retry_warning(retrycount, "Still trying to do this operation");

This increments retrycount every time it is called, but only prints the
log_error() when retrycount is a power of two.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>