ocfs2_controld: Retry CKPT operations indefinitely.
The CKPT service code assumed that TRY_AGAIN was a rare case.  However,
openais returns TRY_AGAIN whenever it's busy doing anything.  So we
can't just retry twice and give up.

It's safe to retry indefinitely in the face of TRY_AGAIN, so we'll do
so.  During initialization, we won't be allowing mounts yet.  During
shutdown, we want to wait for shutdown.  Any time, the process can be
killed.

We're also going to wait indefinitely when we see EEXIST while creating
a checkpoint.  EEXIST means one of two things:

 1) Another daemon is up and running.  This one is just going to sit
    here printing to the log until it's killed or the other one dies.
    This will confuse people; they'll stop the running daemon, but not
    be able to unload the stack.  We have to do this because of the
    second reason.

 2) The daemon was stopped and then immediately restarted.  AIS cleans
    up the checkpoint in a lazy fashion, so there is no guarantee the
    checkpoint is gone by the time the new daemon starts up.  So we can
    get an EEXIST for a little while until AIS gets around to the
    cleanup.  Because scheduling, etc, can take a while, we don't know
    how long that will be.  So we keep retrying.  Eventually, AIS will
    clean up the checkpoint from the daemon that exited and let us
    create our new one.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>