ocfs2_controld: Retry the global checkpoint until it's installed.
The global checkpoint is installed by the first ocfs2_controld to join
the group.  It's read by every subsequent ocfs2_controld to join the
group.  However, if they are all starting at the same time, the
subsequent daemons may try to read the global checkpoint before the
first daemon is finished installing it.

We teach the subsequent daemons to keep retrying until the global
checkpoint is installed.  If the first daemon dies before installing the
checkpoint, the subsequent daemons will give up.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>