fsck.ocfs2: Use the I/O cache.
fsck.ocfs2 travels the filesystem multiple times.  The I/O cache should
make this faster.  Since read-write fsck is only allowed when there are
no other users or mounters of the device, the cache should be safe.

We use two caches.  First, we allocate a cache big enough for all the
journals.  Since we don't know their size at the start, we guess the
default 256MB.  The hope is that we cache the journal blocks on the
first pass when we check their contents and avoid having to re-read them
on the second pass when we replay them.

Once the journals are replayed, we drop this cache and try to allocate a
cache equal to the number of blocks in the filesystem.  This should,
hopefully, keep all of fsck in cache.

We make sure to mlock() our cache, because it's pointless to swap out
cache data; we'd rather just read it from the device.  Now, obviously,
we can't allocate and lock more memory than the system has available.
fsck will keep shrinking the cache size until it gets an allocation.

For the main fsck operation, we don't just get the largest cache
available.  We will need memory for the fsck accounting structures too.
fsck will start with a cache _larger_ than needed.  If this
succeeds, fsck knows that the needed size is safe to allocate.  fsck
will actually use a cache smaller than the largest cache it could get,
ensuring available memory.

Signed-off-by: Joel Becker <joel.becker@oracle.com>