fsck.ocfs2: Pre-fill the I/O cache with metadata.
In pass0, we walk all of the suballocators to verify they look OK.  In
the walk, we read each group descriptor.  Because each group is a linear
hunk of disk, reading the entire group in one slurp is about the same
amount of effort for the disk.  The big problem is the seek, not the
data.  So with almost no impact to pass0, we now pre-fill the I/O cache
will all of our inodes and metadata blocks.

In pass1, this should mean almost everything is in cache if we had a big
enough cache.  If we didn't, oh well.  The worst case is about identical
to the uncached case.

Signed-off-by: Joel Becker <joel.becker@oracle.com>