libocfs2: Catch memalign()s that will abort older glibcs.
Older glibcs (before 2007/07, this includes the glibc in el5) would
abort if __libc_memalign() couldn't allocate the memory.  That's
obviously a bogus behavior, but we have to handle it.

It's simple, though.  We try with malloc() first.  If that succeeds, we
know the memory is there and retry with posix_memalign().

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>