fsck.ocfs2: Pre-cache dirblocks before we go through them.
When we come out of pass 1, o2fsck has a sorted rbtree of dirblock
addresses.  Pass 2 runs that list and checks each dirblock.  However,
it currently reads them one block at a time.

The basic operation of pass 2 is a simple loop that iterates the
dirblocks in block number order.  It passes the dirblock to a callback
that does the checking.  This callback reads the dirblock and the inode
it belongs to.

I tried three caching approaches:

1) Walk the dirblocks, collecting adjacent ones into single I/Os.  Read
   them to pre-fill the cache.  When o2fsck_worth_caching() returns
   false, we know we've filled the cache with dirblocks.  Go ahead and
   process that many of them.  Then go back and read the next hunk of
   dirblocks.  Keep repeating this until all dirblocks are processed.

2) The same as (1), except we pre-cache the inode associated with each
   dirblock as well.

3) A simpler scheme where we just try to read the current dirblock and
   any adjacent ones following it.  Then we process those blocks.  So
   instead of "fill the cache, then process what's in the cache", this
   is "one read, then process what we read".

Approach (1) was the clear winner.  Depending on the cache size, (3) was
either identical or worse than (1).  Approach (2) was just plain worse.
I think this was due to the seek penalty of going off to get the inode
while pre-caching.  Without getting the inode, all our reads are in
ascending order.  Obviously approach (1) has to go get the inode during
the processing phase, but that doesn't impact the pre-cache reads.

Signed-off-by: Joel Becker <joel.becker@oracle.com>