libocfs2: Prevent endian swapping from scribbling over memory.
Some ocfs2 disk structures, such as our btree extent lists, are variable
in size.  They live inside another disk structure, and one of their
fields describes how big they are.  When we read the structure into RAM
and swap it to the local CPU, we rely on that field to make sure we only
swap what is necessary.

However, if that field gets corrupted, our endian swapping code will
happily keep scribbling past the end of the block buffer until it his
the corrupted limit or segfaults.  If you're lucky, it segfaults.

This change modifies all the impacted swapping functions.  They now take
the ocfs2_filesys as an argument so that they know the blocksize.  A new
function, ocfs2_swap_barrier(), is introduced.  It uses the
ocfs2_filesys and the current location in the block buffer to determine
when the swapping code is about to walk off the end.

Some of our list structures can live inside multiple containing
blocks.  Those swap functions require an additional argument, the
address of the containing object.  That way, ocfs2_swap_barrier() knows
what address to start at.

This is ocfs2 bugzilla #1109.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Tao Ma <tao.ma@oracle.com>