fsck.ocfs2: Fix the cluster count if we changed it in pass 0.
At the start of pass 0, we fix the superblock's cluster count to match
the global bitmap.  After we do that, we actually verify the global
bitmap.  If we end up repairing the global bitmap in any way, we never
check to see if its cluster count changed.

This patch checks to see if the cluster count changed during repair.  If
so, the user is prompted to trust the new values.  This triggers a
reinit of the fsck state and a rescan of the global bitmap.  While the
fixed bitmap should be fine, the rescan is necessary to pick up the
clusters used.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>