fsck.ocfs2: Remove the unused ost_bad_inodes bitmap.
We allocate a bitmap of bad inodes.  We test it in pass 2.  But we
never, ever set a bit in it.

It looks to be cloned from e2fsck.  In e2fsck, they don't fix invalid
inode fields in pass 1.  They just mark the inode in the bad map.
Later, in pass2, they test the bad map and fix the fields there.  The
advantage is that they can print a pathname.

That's a major rework for o2fsck, so we're just going to remove the
unused bad map for now.  If someone wants to add it back later, they
can.

Signed-off-by: Joel Becker <joel.becker@oracle.com>