tunefs.ocfs2: Check the bitmap with the correct cluster count.
When tunefs resizes a filesystem, the lowest-level ocfs2_filesys has its
fs_clusters updated, but the master ocfs2_filesys does not.  When tunefs
exits, the bitmap check uses the master ocfs2_filesys.

This adds tunefs_update_fs_clusters(), which stores off any changes to
fs_clusters.  When exiting, this value is used for the bitmap check.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Tao Ma <tao.ma@oracle.com>