tunefs.ocfs2: Don't use the I/O cache in unlocked or online operations.
It's not safe to cache disk blocks when the filesystem may be live
elsewhere.  So if we're running any of the "special" operations, ones
that get the special error codes (SKIPCLUSTER, PERFORM_ONLINE, etc),
we'll run without the cache.

Because ocfs2ne runs the special operations first, then closes the
filesystem and repoens it before trying regular operations, we know that
regular operations will get the cache if they can safely lock the
filesystem.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>