tunefs.ocfs2: Size the cache appropriately for large operations.
This introduces the flag TUNEFS_FLAG_LARGECACHE.  Operations that do not
use this flag get a small I/O cache.  Operations that specify the flag
get a cache as big as they can allocate, up to the size of the
filesystem.  This should speed up operations that have to scan the whole
disk in more than one pass.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>