fsck.ocfs2: Implement Pass 1D's prompt for dup fixes.
For each inode claiming shared clusters, ask about cloning or deleting
them.

Chain allocators are skipped, as they are not safe to clone or delete.
They are very location specific.  If they are sharing with another chain
allocator, pass0 should have handled it.  If they are sharing with a
regular inode, that inode can be cloned or deleted.

The other system files can be cloned but not deleted.

Signed-off-by: Joel Becker <joel.becker@oracle.com>