debugfs.ocfs2: Add a fragmentation score to the 'frag' command.
Fragmentation is often looked at to see the effect on I/O.  But a 100%
fragmented file of 1MB clusters will perform a lot better than a 10%
fragmented file of 4K clusters.

This adds a 'fragmentation score' to the 'frag' command of
debugfs.ocfs2.  The score is simply calculated as the fragmentation
percentage times the number of clusters per megabyte.  Why a megabyte?
Partly because it's our largest cluster size, but also because we know
1MB I/Os are a nice sweet spot.

This gives us an easy way to compare the I/O of files across cluster
sizes.  The higher the score, the worse the I/O.

A file with one extent has a score of 0.  Our theoretical 100%
fragmented file of 1MB clusters has a score of 100.  The 10% fragmented
file of 4K clusters has a score of 2560.  Ouch!  That 10% fragmented
file has an average extent size of 40K.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>