libocfs2: Write leaf extent block in case of record replacement.
In ocfs2_split_extent, when we meet with the situation
of c_split_covers_rec, the old solution just replace the extent
record and forget to write the leaf block. This patch fixes it
and since the root of an extent tree can be anything, so we
leave it to the caller.

Mainline commit is 47be12e4eec84c1846f29af64fe25a396b57a026.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>