Add compress dirs option in fsck.ocfs2 [v3]
This patch adds the option to compress directories (-D) in fsck.
The option compresses the directory entries and consolidates a hole at
the end of the directory block. This consolidated hole increases the
chance of ocfs2_prepare_dir_for_insert (kmp) to find a hole using
ocfs2_find_dir_space_* functions while inserting a new directory entry
in the directory.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Acked-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Tristan Ye <tristan.ye@oracle.com>