inline/xattr: Let ocfs2_file_write handle inline xattr properly.
The kernel now supports both inline data and inline xattr, so
we'd better support them also in user space. But now the user
space libocfs2 still use the old ocfs2_max_inline_data, and
we can easily corrupt a file's inline xattr with following steps:

1. create an empty file by touch.
2. set some xattrs in the file which will resides inside inode.
3. write a user space program and call ocfs2_file_write like this:
	memset(buf, 0, 3800);
        ret = ocfs2_file_write(ci, buf, 3800, 0, &wrote);

In case of bs=4k, the max inline data count without xattr is
3896, so the xattrs we set in step 2 is cleared by ocfs2_file_write.
This patch is try to resolve this bug.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>