fsck.ocfs2: Generic ocfs2_extent_rec check.
unwritten extents actually only exists in the extent tree
for the inode, but we check it for every tree types. And
as for blkno alignment check, refcount tree doesn't need
this type of check.

This patch add a check_leaf_rec_func in extent_info so that the
caller of check_el can initialize it to do its own check for
the leaf ocfs2_extent_rec.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>