libocfs2: Add function for punch hole in refcount tree.
In fsck.ocfs2, sometimes we need to remove the refcount
from the tree, so this patch add a new function named
ocfs2_refcount_punch_hole which will do like this.

We add a new parameter value to ocfs2_decrease_refcount_rec,
so that in case of punching hole, we just need to set
value to -r_refcount and it will work for us.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>