fsck.ocfs2: Store refcounted clusters.
As for the leaf extent record, we generize a function
mark_rec_alloc_func in case they are allocated from the
global_bitmap. As for refcounted clusters, we store them
to a rb-tree so that we can check them later.
For the leaf refcount rec, they are allocated from extent
alloc, so no need to set them.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>