libocfs2: Add function for changing refcount flag.
In fsck.ocfs2, we need the ability to clear/set refcount
flag for some exetent record. So this function adds
a new function named ocfs2_change_refcount_flag.

Note:
It only have the parameter tuple(v_cpos, p_cpos, clusters),
so we have to iterate dinode and xattr to find the dest
extent tree we want to change.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>