fsck.ocfs2: Record refcount tree we have checked.
We check refcount tree if an inode have one. But there
are many inodes sharing the same tree, and we only need
to check once.

This patch adds a rb-tree to store the already-checked
refcount tree so that we can skip them. Also we added
the checked inodes to a list so that it can be used later
by refcounted clusters check later.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>