libo2dlm: Don't drop locks at unlock.
Currently, libo2dlm unlinks files from dlmfs when a lock is unlocked.
This drops the lock in the DLM, losing the contents of the LVB and
removing any lock caching.  While o2dlm holds onto a lock for a few
seconds, no DLM guarantees this will happen.  We don't want that!

Instead, we change o2dlm_unlock() to leave the lock file around in
dlmfs.  This is a truly cached lock.  The calling program can explicitly
drop the lock with o2dlm_drop_lock().  Otherwise, the lock is dropped
when the domain is shut down.

lvb_torture now passes with fsdlm.

Signed-off-by: Joel Becker <joel.becker@oracle.com>