mkfs.ocfs2: Allocate groups to the extent allocator
Extent allocator comes into play when an inode can no longer store all the
data extent pointers in the inode block and needs an indirect block that is
allocated from the extent allocator. This typically happens when the free
space is very fragmented. However, when that is the case, the file system
is also unable to find 4MB contiguous free space that is typically required
to extend this allocator.

To address this issue, we are now preallocating the extent allocator for
all slots and also teaching the file system to steal extents from other
slots when it is unable to extend its own slot allocator.

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>