dx_dirs: mkfs.ocfs2 support
This adds only basic support to mkfs.ocfs2 to write the seed fields used
in directory indexing. The feature string
'indexed-dirs' is already there from a previous patch, so it is possible
to create a new file system with the directory
indexing feature turned on.

Future improvements that are needed:

- mkfs.ocfs2 should create the root directory and orphan dirs as
  indexed.
- possibly also we want to do this for the system dir, but that might
  require some discussion.

[modified the patch for code rebase and cleanup -- Coly Li]

Signed-off-by: Mark Fasheh <mfasheh@suse.com>
Signed-off-by: Coly Li <coly.li@suse.de>