o2image: add interactive mode flag
There is very little danger to the user from taking an Ocfs2 image. We can
write the image out, and if we -ENOSPC the user will see it right away and
can free up space. Thus, there is no "data-integrity" need to query the
user.

The query in o2image isn't like most of our other tools either. o2image only
queries *after* the entire file system is scanned. This means that the user
will have to sit through an expensive operation before being queried to
create the image file.

A user trying to mkfs or tunefs a file system is queried almost immediately,
and always within the reasonable amount of time that an admin would expect a
tool to either finish, continue without interruption or "ask me something".

There is one good reason to query at this moment however - it could be that
the user is low on disk space and would need to free some up before
continuing. So while we want to allow a "fallback query", I don't feel it
should be the default and in fact it does more harm as the default since it
tends to generate confusion in first time users.

This patch changes o2image to proceed with image creation, unless the '-i'
or "interactive mode" switch is given. Users using this switch get exactly
the behavior we have today - once the image size is known, we tell them how
much space it will take so they have a chance to clean up files.

Writes of the image back to a block device are a seperate issue and are not
addressed by this patch. As the code is written today, o2image always
queries in "install" mode (-I).

Signed-off-by: Mark Fasheh <mfasheh@suse.com>
Signed-off-by: Srinivas Eeda <srinivas.eeda@oracle.com>