debugfs.ocfs2: Add the -t and -T options to the "stat" command.
The "stat" command walks the extents or chains of an inode to print
details about its metadata.  This can be expensive when there are a lot
of extents or chains and the user merely wants information from the
inode itself.  The -t and -T options select and deselect traversal of
this metadata respectively.  The -t option remains the default.

Signed-off-by: Joel Becker <joel.becker@oracle.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>