fsck.ocfs2: Change local variable datatype to avoid infinite loop
fsck on large filesystems goes in an infinite loop.
The problem is in verify_bitmap_descs(). i, a local variable is
declared as uint16_t and is compared with
ocfs2_cluster_group_sizes.cgs_cluster_groups which is uint32_t.
When cgs_cluster_groups is greater than 65535, i overflows and wraps
creating an infinite loop of the following:

        for (i = 0, blkno = ost->ost_fs->fs_first_cg_blkno;
             i < cgs.cgs_cluster_groups;
             i++, blkno = i * ocfs2_clusters_to_blocks(ost->ost_fs,
                                                       cgs.cgs_cpg)) {

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Signed-off-by: Mark Fasheh <mfasheh@suse.com>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>