Fix - fsck.ocfs2 attempts to clear dx_dir blocks
If a directory has dx_dir blocks, fsck attempts to clear them in global bitmap.

Steps to reproduce:
Make filesystem with indexed-dirs feature enabled. Make a directory and create
a lot of files in it. Run fsck.ocfs2 -f on this filesystem.

This fix accounts for the dx_leaf nodes and avoids clearing in global bitmap.
The patch reuses the normal file extent functions, which includes code
for refcounting which might not be relevant in this situation.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Signed-off-by: Tao Ma <tao.ma@oracle.com>