bitmap: Add 2 fields to facilitate discontiguous block group.
For a discontiguous block group, we will have some bitmap
region which doesn't start from the bit offset aligned to bytes.
And I don't want to create a new bitmap region which start
from that bit offset because in that case when we read/write
from the original bg->bg_bitmap, we have to shift bits for
every bytes in the region. That is too bad.

So here the solution is: Add a new field named br_bitmap_start
in ocfs2_bitmap_region which indicates the start bit in it.
The bits ahead of it won't be touched by bitmap function now.

With it, we can memcpy the original bitmap easily from the
bg_bitmap and manipulate them.

'br_total_bits' now includes both the useful and useless bits
and another auxiliary paramter br_valid_bits is added to indicate
the really bit length.

Note:
The only left problem is merge_region, but as we don't support
it in chainalloc, we don't need to handle it by now.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>