libocfs2: Let chainalloc handle discontig block group.
chainalloc is used to allocate some bits from a suballocator.
It will use bitmap_region to handle the real work of allocation.
So for contiguous group, create one region for the whole group.
For a discontiguous one, every leaf extent record will become an
individual ocfs2_bitmap_region and be inserted.

Signed-off-by: Tao Ma <tao.ma@oracle.com>
Signed-off-by: Joel Becker <joel.becker@oracle.com>