O2info: Build a main framework for o2info.
This patch tries to build main framework with necessary *.h and *.c
files being generated. it did the option parsing and initialization,
and only works for --help and --version, more operations will be
elaborated in later patches.

To be more detail: all of o2info's operations will be linked to
a task list, and executed one by one later.

Workflow of adding a new operation task for o2info(such as --volinfo)
will be like following:

1. Declare a new extern operation structure.
   extern struct o2info_operation volinfo_op;

2. Add a new option:
   static struct o2info_option volinfo_option = {
        .opt_option     = {
                .name           = "volinfo",
                .val            = 'V',
                .has_arg        = 0,
                .flag           = NULL,
        },
        .opt_help       =
                "-V|--volinfo",
        .opt_handler    = NULL,
        .opt_op         = &volinfo_op,
        .opt_private    = NULL,
};

3. Implement necessary utility func in utils.c.

4. Implement the real running codes for volinfo_op in operation.c.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Acked-by: Sunil Mushran <sunil.mushran@oracle.com>