libo2cb: Validate cluster name and cluster flags in classic_begin_group_join()
In the classic stack with only local heartbeat, classic_begin_group_join() would
always create a heartbeat region that would implicitly check the validity of the
cluster name.

In global heartbeat, heartbeat regions are created before hand. Thus the need
to explicitly check the cluster name.

Also, it ensures the heartbeat mode (encoded in the cluster flags) is also
the same.

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>
Acked-by: Joel Becker <joel.becker@oracle.com>