O2info: Make parts of o2info routines public.
The idea here is to make parts of o2info operations public, a very original
intention is driven by the thoughts of having some o2info functions sharable
with debugfs.ocfs2, so that we'll be able to do 'freefrag'/'freeinode' things
in debugfs as well.

While such kind of o2info funcs were not that friendly/common enough to become
sharable globally. that's the reason why we keep these under o2info in terms of
libo2info instead of libocfs2.

Further patches about 'freefrag' and 'freeinode' of o2info/debugfs.ocfs2 is going
to be posted, being based on this patch.

Signed-off-by: Tristan Ye <tristan.ye@oracle.com>
Acked-by: Sunil Mushran <sunil.mushran@oracle.com>