debugfs.ocfs2: Swap to little-endian before computing metaecc
debugfs.ocfs2 disables metaecc checks while reading the blocks so as to allow
it to show blocks that have incorrect metaecc. Such blocks are shown with their
current and expected metaecc values. 

To do that, it has to compute it manually. But it encounters a problem in that
the metadata blocks have already been swapped to the cpu format in the read
routine. The metaecc compute only works in the little-endian format.

This patch teaches debugfs to detect all metadata blocks and swap their
elements to the little endian format before computing the metaecc.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>