extras: Add utility resize_slotmap.c
resize_slotmap is a utility that resizes the //slot_map file. It does not extend
or truncate the file. It only changes the i_size within the set bounds. The upper
limit is the allocated size (1 cluster) and the minimum space is 2040 bytes
(OCFS2_MAX_SLOTS * sizeof(struct ocfs2_extended_slot)).

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>
Reviewed-by: Joel Becker <jlbec@evilplan.org>