fsck.ocfs2: Fix holes in directories
Calculate the expected offset as the extent tree is traversed and
compare with the offset in the extent record. If the file is not
supposed to have holes, such as directories, prompt and correct it.

Currently no_holes is set for directories only.

I agree this is an ugly hack which bounces variables to the extent
traversal functions, but it gets the job done.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>