fsck.ocfs2: Fix index of directories
In case the i_size of the directory changes, reset the dx_root of the
inode, so it may be later detected for index directory rebuilds. In
pass2, if the filesystem supports directory indexing, rebuild directory
indexes.

I used this method because otherwise we will have to bounce
dirblock_data struct between passes. The dirblock_data is used only in
pass1.

This also helps the patchset - avoiding EROFS on index dir corruptions.
http://oss.oracle.com/pipermail/ocfs2-devel/2011-August/008312.html

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>