fswreck: Add EXTENT_OVERLAP and EXTENT_HOLE
EXTENT_OVERLAP creates a file with extents overlapped, so that there
could be multiple extents for a given extent.

EXTENT_HOLE creates a file with a hole. This is not a corruption.
I added it to check that DIR_HOLE does not correct files with holes.
This can also be used to check sparse files in nosparse filesystem
option.

Signed-off-by: Goldwyn Rodrigues <rgoldwyn@suse.de>
Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>