libocfs2: Increase the buffer size in ocfs2_open_inode_scan()
ocfs2_open_inode_scan() is used by utilities that are looking to read all
the inodes. This patch increases the buffer from one cluster to 4MB to improve
the performance of pass 1 in fsck.

The test was run on a 2TB volume having 15M files. The cache size was 820MB.

Pass 1: Checking inodes and blocks

Before:
  I/O read disk/cache: 64698MB / 66MB, write: 0MB, rate: 17.15MB/s
  Times real: 3775.640s, user: 218.995s, sys: 55.703s

After:
  I/O read disk/cache: 64698MB / 66MB, write: 0MB, rate: 37.28MB/s
  Times real: 1737.086s, user: 193.295s, sys: 25.405s

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>