libocfs2: Read the chain allocator efficiently
As the chain allocators are not linked in an ascending block number order,
reading it is a very seeky affair. This is a huge performance bottleneck in
various tools that have to load the allocators.

This patch adds public function ocfs2_cache_chain_allocator_blocks() in libocfs2
to read the chained groups efficiently and warm the cache with the blocks.

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>