fsck.ocfs2: Add readahead in Pass 2
This patch adds code to readahead directory blocks in pass 2. It reads upto
1024 blocks. Those blocks are then verified after which the utility reads
another 1024 blocks. And so on. This scheme prevents blowing up the cache.

What this does not as yet do is pre-read the directory index blocks and the
inodes specified in the directory block. TODO.

The test was run on a 2TB volume having 15M files. The cache size was 820MB.

Pass 2: Checking directory entries

Before:
  I/O read disk/cache: 3897MB / 131MB, write: 0MB, rate: 2.84MB/s
  Times real: 1416.929s, user: 124.587s, sys: 17.412s

After:
  I/O read disk/cache: 3902MB / 3937MB, write: 0MB, rate: 25.82MB/s
  Times real: 303.600s, user: 95.781s, sys: 13.443s

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>