libocfs2: Improve performance of ocfs2_load_allocator()
ocfs2_load_allocator() calls ocfs2_cache_chain_allocator_blocks()
to preread the chained groups. Its impact on fsck performance is
negligible.

However it is expected to have a big impact on other tools that load
the allocators. Especially mkfs. But for that we have to first bump
up its cache size from a default of 8MB.

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>