o2cluster: New utility to modify the cluster stack
Currently the cluster stack on disk can only be stamped by tunefs.ocfs2.
And that utility only stamps the active cluster stack. While this works
well for pcmk and cman, this is problematic for o2cb as it uses the
ocfs2 volumes as the heartbeat device. This leads us to some chicken
and the egg problems.

o2cluster allows updating the cluster stack to the one specified by
the user. It also allows users to query the active stack and query the
stack stamped on disk.

Signed-off-by: Sunil Mushran <sunil.mushran@oracle.com>